# Improve decode performance