This change reduces the amount of calculations with `*big.Int` (`result.Mul` and `result.Add`) in `Decode` by transforming most of the calculations to simple `int64` ones. This improves performance by something between 200% and 500% on my (amd64) laptop, according to the provided benchmarks.

The inner loop goes through a maximum of 10 iterations, since `var m int64` can only hold `58**10`, but not `58**11`.

The change makes the code harder to read, but since you said in the README you're interested in performance improvements, I'll propose it anyway. I'll let you decide whether you prefer performance or readability.