@akamensky 2021-08-28
Can you please add in the comments the benchmark comparison of current master and your change?

The code looks fine (and passes tests), but it the Github Actions benchmark shows lower numbers than the current master branch.