@codesoap 2021-08-28
Hm, that's interesting. What is the architecture of the GitHub Actions benchmark. I could imagine, that my change doesn't perform as well on 32bit machines, since it uses `int64`...

Without my change (current master):
```
$ go test -bench='Decode*'
goos: openbsd
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Core(TM) i5-5300U CPU @ 2.30GHz
BenchmarkDecode32-2       421952              2750 ns/op          11.64 MB/s
BenchmarkDecode64-2       211719              5574 ns/op          11.48 MB/s
BenchmarkDecode128-2       99670             12245 ns/op          10.45 MB/s
BenchmarkDecode256-2       38865             29905 ns/op           8.56 MB/s
BenchmarkDecode512-2       18048             70169 ns/op           7.30 MB/s
PASS
ok      github.com/akamensky/base58     33.534s
```

With my change:
```
$ go test -bench='Decode*'
goos: openbsd
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Core(TM) i5-5300U CPU @ 2.30GHz
BenchmarkDecode32-2      1529082               776.7 ns/op        41.20 MB/s
BenchmarkDecode64-2       929634              1374 ns/op          46.57 MB/s
BenchmarkDecode128-2      449370              2635 ns/op          48.58 MB/s
BenchmarkDecode256-2      217635              5847 ns/op          43.79 MB/s
BenchmarkDecode512-2       98802             12753 ns/op          40.15 MB/s
PASS
ok      github.com/akamensky/base58     120.751s
```