@akamensky 2021-08-28
Sorry, my bad, I was looking at Encode benchmark results, not Decode, not sure why they are different by ~10%. Could be different host with some additional load, or slower memory.

Not sure if the "checks" tab is visible to you here, the results from Github Actions are:

```
goos: linux
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
BenchmarkEncode32-2    	  281442	      4293 ns/op	   7.45 MB/s
BenchmarkEncode64-2    	  100798	     11140 ns/op	   5.75 MB/s
BenchmarkEncode128-2   	   39648	     29457 ns/op	   4.35 MB/s
BenchmarkEncode256-2   	   12741	     93650 ns/op	   2.73 MB/s
BenchmarkEncode512-2   	    3985	    324960 ns/op	   1.58 MB/s
BenchmarkDecode32-2    	 1791676	       674.4 ns/op	  47.45 MB/s
BenchmarkDecode64-2    	  966186	      1327 ns/op	  48.24 MB/s
BenchmarkDecode128-2   	  477118	      2584 ns/op	  49.54 MB/s
BenchmarkDecode256-2   	  223012	      5389 ns/op	  47.51 MB/s
BenchmarkDecode512-2   	  100052	     12001 ns/op	  42.66 MB/s
PASS
```

Yesterday merge (current master tip):

```
 goos: linux
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Xeon(R) CPU E5-2673 v4 @ 2.30GHz
BenchmarkEncode32-2    	  330958	      3548 ns/op	   9.02 MB/s
BenchmarkEncode64-2    	  137366	      9031 ns/op	   7.09 MB/s
BenchmarkEncode128-2   	   47568	     23909 ns/op	   5.35 MB/s
BenchmarkEncode256-2   	   16416	     74359 ns/op	   3.44 MB/s
BenchmarkEncode512-2   	    4789	    249208 ns/op	   2.05 MB/s
BenchmarkDecode32-2    	  576888	      1922 ns/op	  16.65 MB/s
BenchmarkDecode64-2    	  282474	      4173 ns/op	  15.34 MB/s
BenchmarkDecode128-2   	  136521	      9551 ns/op	  13.40 MB/s
BenchmarkDecode256-2   	   50234	     23449 ns/op	  10.92 MB/s
BenchmarkDecode512-2   	   20923	     61718 ns/op	   8.30 MB/s
PASS
```