@codesoap 2021-08-28
I learned from @mvdan about https://github.com/aclements/perflock. I'm not exactly sure what the tool does, but it somehow makes performance benchmarks less noisy, which helps when comparing performances. I did not use it here either, though...

I'm not sure, however, if it can be enabled for the GitHub Action...